progress on python bin wrappers
