+++ reset_local_run.sh
@@ -1,0 +1,4 @@
+rm -rf $1/chunks/
+rm -rf $1/sgf/
+mkdir $1/chunks
+mkdir $1/sgf
